VERSION 0.5.0.001: Changing splitting criterion from LONG DOUBLE to DOUBLE. R has only precision DOUBLE and using LONG DOUBLE leads to inconsistencies when saving and loading the forest.

+++ src/RFNode.cpp
@@ -35,7 +35,7 @@ void RFNode::setLeafNode(
 
 void RFNode::setSplitNode(
   size_t splitFeature,
-  long double splitValue,
+  double splitValue,
   std::unique_ptr< RFNode > leftChild,
   std::unique_ptr< RFNode > rightChild
 ) {
